Updated testing to be more realistic
Up until now, the test suite that was used to evaluate the LLM-made fix was the test suite that also tested the buggy state. Usually after the human bugfix is introduced, the tests are extended to cover that specific case as well. Now this extended test suite is ran on the LLM-made fix (if the human made changes to test files). Maybe down the road, I may add an option so the LLM tries to generate a test for its own bugfix