hotfix for old poetry commits
Noticed an issue with older commits from repos that use poetry for dependencies: for some commits, the poetry.lock and pyproject.toml files use conflicting versions of dependencies. One of the reasons for poetry would be to keep track of dependency versions for each commit, so the bugs can be reproduced accurately, but if the poetry.lock is not maintained, it does not serve much use (at least to me).